Major updates -all snr scenarios in one dataframe -energy detector alg is now in a dedicated class -removed lr scheduler -added updated notebooks to use the framework -ofdm modulator now considers number of FFT points

+++ src/communications_module/channels/abstract_fading_channel.py
@@ -6,7 +6,7 @@ class AbstractChannel(ABC):
     def __init__(self,
                  discrete_path_delays: list,
                  avg_path_gains: list,
-                 max_doppler_shift: int):
+                 max_doppler_shift: float):
 
         super(AbstractChannel, self).__init__()
         self.discrete_path_delays = discrete_path_delays
