+++ src/communications_module/channels/rayleigh_fading_channel.py
@@ -6,7 +6,7 @@ class RayleighChannel(AbstractChannel):
     def __init__(self,
                  discrete_path_delays: list,
                  avg_path_gains: list,
-                 max_doppler_shift: int):
+                 max_doppler_shift: float):
 
         super().__init__(discrete_path_delays, avg_path_gains, max_doppler_shift)
